build now notifies the user about missing images in the translation

+++ docs/i18n.py
@@ -1,5 +1,7 @@
 import os
 
+from termcolor import colored
+from tabulate import tabulate
 from shutil import copyfile, rmtree, copytree
 
 from conf import LANGUAGES
@@ -7,36 +9,57 @@ from conf import LANGUAGES
 DEFAULT_LANG = 'en'
 PROJET_PATH = os.path.dirname(os.path.realpath(__file__))
 
-def copy_recursive(root, dest, lang=DEFAULT_LANG, path=''):
+def copy_recursive(root, dest, lang=DEFAULT_LANG, path='', silent=False):
     directory = '%s/%s%s' % (root, DEFAULT_LANG, path)
 
+    missing_files = []
+
     for name in os.listdir(directory):
+        if name.startswith('.'):
+            continue
+
         source = '%s/%s' % (directory, name)
         target = '%s%s/%s' % (dest, path, name)
 
         source_i18n = '%s/%s%s/%s' % (root, lang, path, name)
 
-        if(os.path.exists(source_i18n)):
+        if os.path.exists(source_i18n):
             source = source_i18n
+        else:
+            if not os.path.isdir(source):
+                missing_files.append([path or '/', name])
 
         if os.path.isfile(source):
+            if silent:
+                continue
+
             if not os.path.exists(os.path.dirname(target)):
                 os.makedirs(os.path.dirname(target))
 
             copyfile(source, target)
         else:
-            copy_recursive(root, dest, lang, '%s/%s' % (path, name))
+            missing_files = missing_files + copy_recursive(root, dest, lang=lang, path='%s/%s' % (path, name), silent=silent)
 
-def copy_images(lang=DEFAULT_LANG, path=PROJET_PATH):
-    if os.path.exists('%s/_static/images' % path):
-        rmtree('%s/_static/images' % path)
+    return missing_files
 
-    os.makedirs('%s/_static/images' % path)
+def copy_images(lang=DEFAULT_LANG, path=PROJET_PATH, silent=False):
+    if not silent:
+        if os.path.exists('%s/_static/images' % path):
+            rmtree('%s/_static/images' % path)
+
+        os.makedirs('%s/_static/images' % path)
 
     root = '%s/_static/_images' % path
     destination = '%s/_static/images' % path
 
-    copy_recursive(root, destination, lang=lang)
+    missing_files = copy_recursive(root, destination, lang=lang, silent=silent)
+
+    if missing_files:
+        missing_files = sorted(missing_files, key=lambda x: x[0])
+
+        print colored('\nNot all images have translations!', 'red')
+        print colored('The following files still need to be translated to \"%s\"\n' % lang, 'yellow')
+        print tabulate(missing_files, headers=['Folder', 'File'])
 
 def copy_sources(dest, origin=PROJET_PATH):
     BLACKLIST = LANGUAGES + [
